QCBase

Documentation for QCBase.

QCBase.Atom

QCBase.MOCluster

QCBase.Molecule

Base.length

QCBase.dim_tot

QCBase.dim_tot

QCBase.possible_focksectors

QCBase.Atom — Type

id::Integer             #index of atom in the molecule
symbol::String          #Atomic ID (E.g. H, He, ...)
xyz::Array{Float64,1}   #list of XYZ coordinates

Simply an Atom

QCBase.MOCluster — Type

idx::Int16
orb_list::Vector{Int16}

QCBase.Molecule — Type

charge::Integer         #overall charge on molecule
multiplicity::Integer   #2S+1
atoms::Vector{Atom}     #Vector of `Atoms`
basis::String           #Basis set

Molecule essentially as a Vector of atoms, number of electrons and basis set

Base.length — Method

length(c::MOCluster)

Return number of orbitals in MOCluster

QCBase.dim_tot — Method

dim_tot(c::MOCluster, na, nb)

Return dimension of hilbert space spanned by number of orbitals in MOCluster with na and nb number of alpha/beta electrons.

QCBase.dim_tot — Method

dim_tot(c::MOCluster)

Return dimension of hilbert space spanned by number of orbitals in MOCluster. This is all sectors

QCBase.possible_focksectors — Method

possible_focksectors(c::MOCluster, delta_elec=nothing)

Get list of possible fock spaces accessible to the MOCluster

delta_elec::Vector{Int}: (nα, nβ, Δ) restricts fock spaces to: (nα,nβ) ± Δ electron transitions